Split out safeLink
