+++ client/lib/post-normalizer/utils/safe-link.js
@@ -1,0 +1,14 @@
+export const safeLinkRe = /^https?:\/\//;
+
+/**
+ * Only accept links that start with http or https. Reject others.
+ *
+ * @param {string} link the link to check
+ * @returns {string|undefined} the safe link or undefined
+ */
+export function safeLink( link ) {
+	if ( safeLinkRe.test( link ) ) {
+		return link;
+	}
+	return undefined;
+}
